Update-Finish message bug

+++ web/index.php
@@ -105,9 +105,11 @@
 
 
 	$waittime = 0;
-	if (isset($_POST['action']) || isLogin() && $_SESSION['benutzer']['typ'] == "admin" && file_exists("../data/algorithmus.pid") && !isRunning(file_get_contents("../data/algorithmus.pid"))) {
+	if (isset($_POST['action']) || isLogin() && $_SESSION['benutzer']['typ'] == "admin" && (
+		!file_exists("../data/cleanup.lock") && file_exists("../data/algorithmus.pid") && !isRunning(file_get_contents("../data/algorithmus.pid")) ||
+		!file_exists("../data/update.lock") && file_exists("../data/update.pid") && !isRunning(file_get_contents("../data/update.pid")))) {
 		// cleanup des Zuteilungsalgorithmus
-		if (!file_exists("../data/cleanup.lock") && file_exists("../data/algorithmus.pid") && !isRunning(file_get_contents("../data/algorithmus.pid"))) {
+		if (isLogin() && $_SESSION['benutzer']['typ'] == "admin" && !file_exists("../data/cleanup.lock") && file_exists("../data/algorithmus.pid") && !isRunning(file_get_contents("../data/algorithmus.pid"))) {
 			$fh = fopen("../data/cleanup.lock", "w");
 			fclose($fh);
 		  unlink("../data/algorithmus.pid");
@@ -129,7 +131,7 @@
 			$waittime = 2;
 			unlink("../data/cleanup.lock");
 		}
-		elseif (!file_exists("../data/update.lock") && file_exists("../data/update.pid") && !isRunning(file_get_contents("../data/update.pid"))) {
+		elseif (isLogin() && $_SESSION['benutzer']['typ'] == "admin" && !file_exists("../data/update.lock") && file_exists("../data/update.pid") && !isRunning(file_get_contents("../data/update.pid"))) {
 		  $fh = fopen("../data/update.lock", "w");
 		  fclose($fh);
 		  unlink("../data/update.pid");
